Move around IBC event traits
