+++ crates/relayer-framework/src/base/chain/traits/mod.rs
@@ -1,6 +1,5 @@
 pub mod chain_id;
 pub mod event_subscription;
-pub mod ibc_event;
 pub mod message;
 pub mod message_sender;
 pub mod queries;
