+++ crates/relayer-framework/src/base/chain/traits/queries/write_acknowledgement.rs
@@ -1,6 +1,6 @@
 use async_trait::async_trait;
 
-use crate::base::chain::traits::ibc_event::HasWriteAcknowledgementEvent;
+use crate::base::chain::traits::types::ibc_events::write_ack::HasWriteAcknowledgementEvent;
 use crate::base::chain::traits::types::packet::HasIbcPacketTypes;
 use crate::std_prelude::*;
