+++ crates/relayer-framework/src/base/chain/traits/types/ibc_events/write_ack.rs
@@ -4,26 +4,6 @@ use crate::base::core::traits::sync::Async;
 /**
    Indicates that a chain context's
    [`Event`](crate::base::chain::traits::types::event::HasEventType::Event)
-   type contains a [`SendPacketEvent`](Self::SendPacketEvent) variant.
-*/
-pub trait HasSendPacketEvent<Counterparty>: HasIbcPacketTypes<Counterparty>
-where
-    Counterparty: HasIbcPacketTypes<
-        Self,
-        IncomingPacket = Self::OutgoingPacket,
-        OutgoingPacket = Self::IncomingPacket,
-    >,
-{
-    type SendPacketEvent: Async;
-
-    fn try_extract_send_packet_event(event: &Self::Event) -> Option<Self::SendPacketEvent>;
-
-    fn extract_packet_from_send_packet_event(event: &Self::SendPacketEvent)
-        -> Self::OutgoingPacket;
-}
-/**
-   Indicates that a chain context's
-   [`Event`](crate::base::chain::traits::types::event::HasEventType::Event)
    type contains a [`WriteAcknowledgementEvent`](Self::WriteAcknowledgementEvent) variant.
 */
 pub trait HasWriteAcknowledgementEvent<Counterparty>: HasIbcPacketTypes<Counterparty>
