+++ crates/relayer-framework/src/base/chain/types/aliases.rs
@@ -2,11 +2,11 @@ use alloc::sync::Arc;
 use core::pin::Pin;
 use futures::stream::Stream;
 
-use crate::base::chain::traits::ibc_event::HasWriteAcknowledgementEvent;
 use crate::base::chain::traits::types::chain::HasChainTypes;
 use crate::base::chain::traits::types::event::HasEventType;
 use crate::base::chain::traits::types::height::HasHeightType;
 use crate::base::chain::traits::types::ibc::HasIbcChainTypes;
+use crate::base::chain::traits::types::ibc_events::write_ack::HasWriteAcknowledgementEvent;
 use crate::base::chain::traits::types::message::HasMessageType;
 use crate::base::runtime::traits::runtime::HasRuntime;
 use crate::base::runtime::traits::subscription::Subscription;
