+++ crates/relayer-framework/src/base/relay/impls/packet_relayers/receive/base_receive_packet.rs
@@ -1,7 +1,7 @@
 use async_trait::async_trait;
 
-use crate::base::chain::traits::ibc_event::HasWriteAcknowledgementEvent;
 use crate::base::chain::traits::message::receive_packet::CanBuildReceivePacketMessage;
+use crate::base::chain::traits::types::ibc_events::write_ack::HasWriteAcknowledgementEvent;
 use crate::base::chain::types::aliases::Height;
 use crate::base::relay::traits::ibc_message_sender::{CanSendIbcMessages, IbcMessageSenderExt};
 use crate::base::relay::traits::packet_relayers::receive_packet::ReceivePacketRelayer;
